remove old transaction_search code
