+++ app/inputs/transaction_chosen_input.rb
@@ -1,29 +1,31 @@
-# This class is deprecated in favor of TransactionChosen2Input. This class
-# is still being used by many of the TransactionSearch controller concern views.
-# These should eventually be replaced by TransactionSearch::Searcher (see FacilityAccountsReconciliationController).
-class TransactionChosenInput < SimpleForm::Inputs::Base # CollectionSelectInput
+class TransactionChosen2Input < SimpleForm::Inputs::CollectionInput
 
   disable :required
 
-  def input(_wrapper_options)
-    options[:label_method] ||= :name
-    options[:value_method] ||= :id
+  def input(wrapper_options = nil)
+    label_method, value_method = detect_collection_methods
+
+    merged_input_options = merge_wrapper_options(input_html_options, wrapper_options)
 
     # By calling `to_a` here, it fetches the AR::Relation into objects. Since that will
     # happen later anyways, by doing it up front, we prevent extra `count(*)` queries,
     # which were slowing down the page.
-    collection_size = collection_items.to_a.size
+    collection_size = collection.to_a.size
 
-    select_options = { :multiple => true, :"data-placeholder" => placeholder_label }
+    merged_input_options[:multiple] = true
+    merged_input_options["data-placeholder"] = placeholder_label
 
-    # If there is only one possible value, then we want to show it, and not allow
-    # selection, but only if it's not a nullable field
+    # If there is only one possible value, then we want to show it as selected,
+    # but we don't want to allow adding/removing it.
     if collection_size == 1 && !options[:allow_blank]
-      search_fields[attribute_name] = [collection_items.first.send(options[:value_method].to_sym)]
-      select_options[:disabled] = :disabled
+      input_options[:selected] = [collection.first.public_send(value_method)]
+      merged_input_options[:disabled] = :disabled
     end
 
-    template.select_tag(attribute_name, option_data, select_options).html_safe
+    option_data = option_elems(label_method, value_method)
+    selected = input_options[:selected] || object.public_send(attribute_name)
+    opts = template.options_for_select option_data, selected: selected
+    template.select_tag("#{object.class.model_name.param_key}[#{attribute_name}]", opts, merged_input_options)
   end
 
   def label(wrapper_options)
@@ -33,16 +35,30 @@ class TransactionChosenInput < SimpleForm::Inputs::Base # CollectionSelectInput
 
   private
 
-  def collection_items
-    template.instance_variable_get("@search_options").try(:[], attribute_name)
+  # Pass an array if you want arguments, [:full_name, suspended_label: true]
+  def option_elems(label_method, value_method)
+    collection.map do |i|
+      [
+        i.public_send(*label_method),
+        i.public_send(*value_method),
+        data_for_item(i),
+      ]
+    end
+  end
+
+  def data_for_item(item)
+    data_proc = options[:data_attrs] || ->(_i) { {} }
+    dataify(data_proc.call(item))
+  end
+
+  def dataify(hash)
+    Hash[
+      hash.map { |k, v| ["data-#{k.to_s.dasherize}", v] }
+    ]
   end
 
   def model_label
     attribute_class.model_name.human(count: 2)
-  end
-
-  def search_fields
-    template.instance_variable_get("@search_fields")
   end
 
   def attribute_class
@@ -51,14 +67,6 @@ class TransactionChosenInput < SimpleForm::Inputs::Base # CollectionSelectInput
 
   def placeholder_label
     "Select #{options[:label].downcase}"
-  end
-
-  def option_data
-    if options[:data_method]
-      template.send(options[:data_method], collection_items, search_fields[attribute_name])
-    else
-      template.options_from_collection_for_select(collection_items, options[:value_method], options[:label_method], search_fields[attribute_name])
-    end
   end
 
 end
