+++ app/views/facilities/movable_transactions.html.haml
@@ -7,7 +7,7 @@
 
 - @search_fields = params
 = content_for :top_block do
-  = render "shared/transactions/top2", tab: "movable_transactions"
+  = render "shared/transactions/top", tab: "movable_transactions"
 
 - if @order_details.present?
   = render "shared/transactions/table"
