+++ app/views/facility_notifications/in_review.html.haml
@@ -5,7 +5,7 @@
 = content_for :h2 do
   %h2= t(".head")
 = content_for :top_block do
-  = render "shared/transactions/top2", tab: "in_review"
+  = render "shared/transactions/top", tab: "in_review"
 
 - if @order_details.any?
   = content_for :action_instructions do
