+++ app/views/facility_reservations/index.html.haml
@@ -7,7 +7,7 @@
   = render partial: "admin/shared/tabnav_reservation"
 
 .span12
-  = render "shared/transactions/search2"
+  = render "shared/transactions/search"
 
 .span12
   - if @order_details.any?
