+++ app/views/facility_reservations/show_problems.html.haml
@@ -8,7 +8,7 @@
     locals: { secondary_tab: "problem" }
 
 .span12
-  = render "shared/transactions/search2"
+  = render "shared/transactions/search"
 
 - if @order_details.any?
   .span12
