+++ app/views/shared/transactions/_search.html.haml
@@ -1,22 +1,17 @@
--# This view, which relies on the TransactionSearch controller/concern is
--# deprecated in favor of _search2 and TransactionSearch::Searcher. See FacilityAccountsReconciliationController
--# as an example of the more modern method.
-= modelless_form_for url: url_for(params), method: :get, html: { class: "search_form" }, defaults: { required: false } do |f|
+= simple_form_for @search_form, url: url_for, method: :get, html: { class: "search_form" }, defaults: { required: false } do |f|
   .row
     %fieldset.span6#search
-      = f.input :facilities, as: :transaction_chosen
-      = f.input :accounts, label_method: "account_list_item", as: :transaction_chosen unless @account
-      = f.input :products, as: :transaction_chosen, data_method: :product_options
-      = f.input :account_owners, label: Account.human_attribute_name(:owner).pluralize, label_method: :full_name, as: :transaction_chosen unless @account
-      = f.input :order_statuses, data_method: :order_statuses_options, as: :transaction_chosen
-      = render_view_hook("end_of_first_column", f: f)
-    %fieldset.span2#calendar_filter
-      %br
-      = f.input "date_range[field]", collection: @search_options[:date_range], selected: @date_range_field, label: false
-      = f.input "date_range[start]", input_html: { value: @search_fields[:date_range].try(:[], :start).to_s.gsub("-","/"), class: ["datepicker", "in_past"] }, label: t("reports.fields.date_start")
-      = f.input "date_range[end]", input_html: { value: @search_fields[:date_range].try(:[], :end).to_s.gsub("-","/"), class: ["datepicker", "in_past"] }, label: t("reports.fields.date_end")
+      - @search.options.reject(&:multipart?).each do |searcher|
+        = f.input searcher.key, as: :transaction_chosen2, collection: searcher.options, label: searcher.label, label_method: searcher.label_method, data_attrs: searcher.method(:data_attrs), input_html: { id: searcher.key }
 
-    .submit_button.span6
-      = f.input :email, as: :hidden, input_html: { value: current_user.email }, disabled: true
-      = f.input :format, as: :hidden, input_html: { value: params[:format] }, disabled: true
+    %fieldset.span2#calendar_filter
+      - if @search.options.map(&:key).include?("date_ranges")
+        %br
+        = f.input :date_range_field, collection: TransactionSearch::DateRangeSearcher.options(only: @search_form.allowed_date_fields), label: false, include_blank: false
+        = f.input :date_range_start, input_html: {class: ["datepicker", "in_past"] }, label: t("reports.fields.date_start")
+        = f.input :date_range_end, input_html: { class: ["datepicker", "in_past"] }, label: t("reports.fields.date_end")
+
+    .submit_button.span12
+      = hidden_field_tag :email, current_user.email, disabled: true
+      = hidden_field_tag :format, params[:format], disabled: true
       = f.submit t("shared.filter"), class: "btn float-left"
