--- app/views/shared/transactions/_search2.html.haml
@@ -1,19 +1,0 @@
--# This view, replaces _search, which relies on the TransactionSearch controller/concern.
--# This view, along with TransactionSearch::Searcher is the more modern method. See FacilityAccountsReconciliationController
-= simple_form_for @search_form, url: url_for, method: :get, html: { class: "search_form" }, defaults: { required: false } do |f|
-  .row
-    %fieldset.span6#search
-      - @search.options.reject(&:multipart?).each do |searcher|
-        = f.input searcher.key, as: :transaction_chosen2, collection: searcher.options, label: searcher.label, label_method: searcher.label_method, data_attrs: searcher.method(:data_attrs), input_html: { id: searcher.key }
-
-    %fieldset.span2#calendar_filter
-      - if @search.options.map(&:key).include?("date_ranges")
-        %br
-        = f.input :date_range_field, collection: TransactionSearch::DateRangeSearcher.options(only: @search_form.allowed_date_fields), label: false, include_blank: false
-        = f.input :date_range_start, input_html: {class: ["datepicker", "in_past"] }, label: t("reports.fields.date_start")
-        = f.input :date_range_end, input_html: { class: ["datepicker", "in_past"] }, label: t("reports.fields.date_end")
-
-    .submit_button.span12
-      = hidden_field_tag :email, current_user.email, disabled: true
-      = hidden_field_tag :format, params[:format], disabled: true
-      = f.submit t("shared.filter"), class: "btn float-left"
