--- app/views/shared/transactions/_top2.html.haml
@@ -1,6 +1,0 @@
-.row#header_billing
-  .span3
-    #sidebar
-      = render "admin/shared/sidenav_billing", sidenav_tab: tab
-  .span9
-    = render "shared/transactions/search2"
